requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
google-auth==2.43.0
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
